174-FE: Move `HAPPENED_AT_TIMEFRAME` to constants
For consistency with the other enums.
