+++ Okane.Client/src/features/financeRecords/types/searchFilters.ts
@@ -1,16 +1,9 @@
 // External
+import { HAPPENED_AT_TIMEFRAME } from '@features/financeRecords/constants/searchFilters'
 import { COMPARISON_OPERATOR, SORT_DIRECTION } from '@shared/constants/search'
 
 import { type FINANCE_RECORD_TYPE } from '@features/financeRecords/constants/saveFinanceRecord'
 import { type Tag } from '@shared/types/tag'
-
-// This enum needs to be here and not in constants to prevent an issue due to circular imports.
-export enum HAPPENED_AT_TIMEFRAME {
-  CUSTOM = 'custom',
-  PAST_30_DAYS = 'past30Days',
-  THIS_MONTH = 'thisMonth',
-  THIS_YEAR = 'thisYear',
-}
 
 export type FinanceRecordSearchFilters = {
   description: string
